WIP: Initial commit of performance test and major refactor
