+++ .jshintrc
@@ -1,0 +1,70 @@
+{
+    "maxerr"        : 50,
+
+    "bitwise"       : true,
+    "camelcase"     : true,
+    "curly"         : false,
+    "eqeqeq"        : true,
+    "forin"         : false,
+    "freeze"        : true,
+    "immed"         : true,
+    "latedef"       : true,
+    "newcap"        : true,
+    "noarg"         : true,
+    "noempty"       : true,
+    "nonbsp"        : true,
+    "nonew"         : false,
+    "plusplus"      : false,
+    "quotmark"      : false,
+    "undef"         : true,
+    "unused"        : "strict",
+    "maxparams"     : 5,
+    "maxdepth"      : 4,
+    "maxstatements" : false,
+    "maxcomplexity" : false,
+    "maxlen"        : 80,
+    "varstmt"       : false,
+    "asi"           : false,
+    "boss"          : false,
+    "debug"         : true,
+    "eqnull"        : false,
+    "moz"           : false,
+    "evil"          : false,
+    "expr"          : false,
+    "funcscope"     : false,
+    "iterator"      : false,
+    "lastsemic"     : false,
+    "laxbreak"      : false,
+    "laxcomma"      : false,
+    "loopfunc"      : false,
+    "multistr"      : false,
+    "noyield"       : false,
+    "notypeof"      : false,
+    "proto"         : false,
+    "scripturl"     : false,
+    "shadow"        : false,
+    "sub"           : false,
+    "supernew"      : false,
+    "validthis"     : false,
+
+    "browser"       : false,
+    "browserify"    : false,
+    "couch"         : false,
+    "devel"         : true,
+    "dojo"          : false,
+    "jasmine"       : false,
+    "jquery"        : false,
+    "mocha"         : true,
+    "mootools"      : false,
+    "node"          : true,
+    "nonstandard"   : false,
+    "phantom"       : false,
+    "prototypejs"   : false,
+    "qunit"         : false,
+    "rhino"         : false,
+    "shelljs"       : false,
+    "typed"         : false,
+    "worker"        : false,
+    "wsh"           : false,
+    "yui"           : false
+}
